Break logic out into separate file
Keeping the logic in functions in a separate file will avoid duplication and confusion when working between the standard and lite versions of Dice Maiden
